Width height edit date (#46)
* [#41] Fix the last_edited and edited_by fields

The last_edited field was not being pulled from the database. We are now
pulling it correctly, and added the edited_by field as well.

* [#42] Plumbbed through width and height into sheet

Sheet now takes the canvas width and height in mm. This will
then calculate the correct points for the pdf surface to render
the svg onto.

* [bug] fixed unit conversion, should have been mul not div

* [codereview] removed extra commented code.