Fixed issue with caching bounds
Bounds jobs should not be check to see if it was already submitted. Diffcult to know if it was an excat match as there are not enough data store in in postgres jobs db at current. Plus, it does not really benifite the system as much as MDGID where are a set coordinate system.

Disabled check for bounds based job requests.