[issue #115] Allow sheets to have multiple styles
The config now allows one to define all the styles that will be used
and assign them a name and description in addition to the location.

Then each sheet in the config can use the `styles` attribute to define
the order and which styles defined in the global section will be used
for this sheet. The default style is always the first style defined,
in each list.

If a sheet does not define a set of styles to use, it will inheriate
the list of global styles.

A new field has been added to the jobs js called 'style_name' that can
be used to indicate which style the sheet should use.

In the CLI styles defined in the config can be listed. Also, sheet info
now include the list of style avilable and their descriptions.

There is a new SQL that needs to be applied before this binary can be
deployed. This change will not effect the old version. This version
is also compatiable with the old config file. New Styles will be
generated for each sheet using the old `style` option. If `style` and
`styles` are both found in a sheet configuration only the `styles` will
be used. In addition no warning will be given, unlike if only `style` is
found.

fixes #115