For Flipped row|col calulations off by one.
The row and col checks for the LabelForRow was not
checking to see if the row was strickly smaller then len(row),
causing the inverse to be off by one. Leading to not be able to get
the correct label for the last entry.

Col was doing this as well, but since it does not label things from
right to left, issue did not show up.